Update hook/form for `Connect` to be similar to `ProposalNew`
